refactor(database): Remove ktx imports
