+++ database/app/src/main/java/com/google/firebase/referencecode/database/kotlin/EmulatorSuite.kt
@@ -1,8 +1,8 @@
 package com.google.firebase.referencecode.database.kotlin
 
 import com.google.firebase.database.FirebaseDatabase
-import com.google.firebase.database.ktx.database
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.database.database
+import com.google.firebase.Firebase
 
 class EmulatorSuite {
