+++ database/app/src/main/java/com/google/firebase/referencecode/database/kotlin/MainActivity.kt
@@ -6,9 +6,9 @@ import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.ValueEventListener
-import com.google.firebase.database.ktx.database
-import com.google.firebase.database.ktx.getValue
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.database.database
+import com.google.firebase.database.getValue
+import com.google.firebase.Firebase
 import com.google.firebase.referencecode.database.R
 
 abstract class MainActivity : AppCompatActivity() {
